Move Azure disk-ID parsers to standalone cloudutil sub-package
extractAzureResourceGroup and extractAzureDiskName are pure string
functions but their tests lived in package e2e, requiring CLOUD_PROVIDER
and TEST_PROVISION_FILE just to run TestMain. Move them to
test/e2e/cloudutil/ so they can be tested independently with
`go test ./test/e2e/cloudutil/` and no external setup.

Also fix README docs that incorrectly stated TEST_PROVISION_FILE "can be
empty / can point to any file" — TestMain parses it as TOML, so it must
be valid TOML or an empty file.
